Add a few new preserve-3d hit-testing tests for currently broken behaviour. https://bugs.webkit.org/show_bug.cgi?id=248281 <rdar://problem/102632565>
Reviewed by Simon Fraser.

Neither of these tests should change behaviour with interop, but we can fix them at the same
time.

* LayoutTests/platform/gtk/transforms/3d/point-mapping/3d-point-mapping-2-expected.txt:
* LayoutTests/platform/ios/transforms/3d/point-mapping/3d-point-mapping-2-expected.txt:
* LayoutTests/platform/mac/transforms/3d/point-mapping/3d-point-mapping-2-expected.txt:
* LayoutTests/transforms/3d/point-mapping/3d-point-mapping-2.html:

This adds a new subtest check for hitting the parent box3 element in the area where it's
drawn on top of the child. We're currently failing this, as we hit-test box4 despite it
being rendered behind.

* LayoutTests/transforms/3d/preserve-3d-not-applied-to-ancestors-expected.txt: Added.
* LayoutTests/transforms/3d/preserve-3d-not-applied-to-ancestors.html: Added.

This test has a transformed ancestor of the preserve-3d element, which shouldn't be
participating in transform accumulation (and rendering already shows that it isn't).
The bounding client rect should match, and it currently doesn't.

Canonical link: https://commits.webkit.org/257091@main