Fix potential crash under IntersectionObserver::disconnect() https://bugs.webkit.org/show_bug.cgi?id=248111 rdar://100355921
Reviewed by Jonathan Bedard and Ryosuke Niwa.

Make sure we protect the intersection observers and resize observers before
calling disconnect() on them in Document::commonTeardown().

This is a speculative fix to address the crash in the radar, which I was
unable to reproduce.

* LayoutTests/fast/dom/lazy-loading-iframe-destruction-crash-expected.txt: Added.
* LayoutTests/fast/dom/lazy-loading-iframe-destruction-crash.html: Added.
Include test from the radar, even though it didn't reproduce the issue for me.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::commonTeardown):

Canonical link: https://commits.webkit.org/252432.841@safari-7614-branch