Closing ARQL results in an empty view with a spinner https://bugs.webkit.org/show_bug.cgi?id=248538 rdar://102379447
Reviewed by Antoine Quint.

ARQL compares the URL we give it at the start of a download to the one
we give it at the end. The latter one has the fragment identifier appended,
and unfortunately ARQL does a string comparison. This causes it to think
it is opening two different files, resulting in a spinner spinning forever.

* Source/WebKit/UIProcess/Cocoa/SystemPreviewControllerCocoa.mm: Remove the
fragment identifier before sending it to ARQL.

Canonical link: https://commits.webkit.org/257188@main