Always make ::backdrop renderers the child of the RenderView except for <dialog> https://bugs.webkit.org/show_bug.cgi?id=248569 rdar://102833504
Reviewed by NOBODY (OOPS!).

Some renderers don't expect extra children like RenderVideo, which causes the ::backdrop to not be laid out. We could fix all renderers to lay out children, but that's probably unnecessary complexity and it's likely to miss certain cases.

So let's append all ::backdrop to the RenderView whenever we can.

* Source/WebCore/rendering/updating/RenderTreeUpdaterGeneratedContent.cpp:
(WebCore::RenderTreeUpdater::GeneratedContent::updateBackdropRenderer):