Fix constant timeouts in transition tests https://bugs.webkit.org/show_bug.cgi?id=248449
Reviewed by Darin Adler.

In 257041@main I changed the computed getPropertyCSSValue() for the
border-radius longhands. Previously it was either a CSSPrimitiveValue
holding a length, or a CSSValueList with 2 CSSPrimitiveValue lengths.
To make it consistent with the parser, I made it always return a
CSSPrimitiveValue holding a Pair of CSSPrimitiveValue lengths.

This affected transition-test-helpers.js, which didn't have logic for
the border-radius longhands, so it was falling back to getFloatValue().
This would have already failed in the CSSValueList case, but it went
undetected until I used the CSSPrimitiveValue pair.

Then getFloatValue() started throwing an exception, which would prevent
the tests from completing. Hence a timeout.

This patch adds custom logic for the border-radius longhands, and adds
a try...catch so that exceptions don't cause a timeout.

* LayoutTests/TestExpectations:
* LayoutTests/transitions/resources/transition-test-helpers.js:

Canonical link: https://commits.webkit.org/257208@main