[EME][Thunder] Prune cached sessions https://bugs.webkit.org/show_bug.cgi?id=248609
Reviewed by NOBODY (OOPS!).

The framework should inform us of keys that are no longer available but it does not so we keep caching them when coming.
The problem seems to be that DRM systems run out of sessions. To avoid that we are putting in place a mechanism to allow
DRM backends to un-cache sessions we are are going to regularly check if keys are still in use or not. If they are in
use, we don't un-cache them but if they are not, we can un-cache that framework session.

We are not informing the JS app about those keys being unavailable because theoretically they still are available and we
should be able to ask the framework for a session for a certain key id again.

* Source/WebCore/platform/encryptedmedia/CDMProxy.cpp:
(WebCore::KeyHandle::takeValueIfDifferent):
(WebCore::KeyStore::KeyStore):
(WebCore::KeyStore::~KeyStore):
(WebCore::KeyStore::pruneKeysIfNeeded):
* Source/WebCore/platform/encryptedmedia/CDMProxy.h:
(WebCore::KeyHandle::value const):
(WebCore::KeyHandle::value):
(WebCore::KeyHandle::pruneIfNeeded):
(WebCore::KeyHandle::markUsed const):
(WebCore::KeyHandle::KeyHandle):
* Source/WebCore/platform/graphics/gstreamer/eme/CDMThunder.cpp:
(WebCore::KeyHandleThunder::KeyHandleThunder):
(WebCore::KeyHandleThunder::pruneIfNeeded):
(WebCore::KeyHandleThunder::markUsed const):
(WebCore::CDMInstanceSessionThunder::keyUpdatedCallback):
* Source/WebCore/platform/graphics/gstreamer/eme/CDMThunder.h:
(WebCore::KeyHandleThunder::create):