Calling StylePropertyMap.set() with a `calc()` ends up dropping the calc https://bugs.webkit.org/show_bug.cgi?id=248559
Reviewed by Geoffrey Garen.

When calling StylePropertyMap.set() with a CSSMathValue, we would call
CSSMathValue::toCSSValue() to convert it to a CSSValue. We would expect to get
a CSSCalcValue as a result. However, our implementation was trying to resolve
the `calc()` and would return a CSSPrimitiveValue instead.

* LayoutTests/fast/css/css-typed-om/style-property-map-set-CSSMathSum-value-expected.txt: Added.
* LayoutTests/fast/css/css-typed-om/style-property-map-set-CSSMathSum-value.html: Added.
* Source/WebCore/css/typedom/numeric/CSSMathValue.h:

Canonical link: https://commits.webkit.org/257291@main