[macOS / iOS] Limit WPT layout tests to mac-wk2 configuration in CI https://bugs.webkit.org/show_bug.cgi?id=248671 rdar://102913568
Reviewed by NOBODY (OOPS!).

WPT layout tests make up for over half of the overall layout test runtime on macOS and iOS bots.
This is affecting throughput in CI, and incurs a large cost for bot watchers to maintain.
In an attempt to alleviate those issues, add a new TestExpectations file used by CI to restrict
these tests to the mac-wk2 configuration. Developers will still be able to run these tests
locally without any change in workflow, and modifying the new TestExpectations file will
force tests to run in EWS if needed for testing.

* LayoutTests/continuous-integration/TestExpectations: Added.
* Tools/CISupport/build-webkit-org/config.json:
* Tools/CISupport/ews-build/config.json: