Regression(255052@main) Unable to log into Marcus.com on iOS https://bugs.webkit.org/show_bug.cgi?id=248798 rdar://102959860
Reviewed by Brent Fulgham.

Since 255052@main, login on marcus.com fails on iOS and complains about being
outside the USA. The same thing happened by we dropped `window.showModalDialog()`.
The site's login seems to rely on which properties are present on the `window`
and `navigator` object and what their values are.

In this case, the log in fails because `navigator.plugins` and
`navigator.mimeTypes` now returns the values hard-coded in the HTML
specification instead of being empty arrays.

This is a site issue so I am adding a quirk for now and contacted evangelism.

* Source/WebCore/page/Navigator.cpp:
(WebCore::Navigator::initializePluginAndMimeTypeArrays):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldNavigatorPluginsBeEmpty const):
* Source/WebCore/page/Quirks.h:

Canonical link: https://commits.webkit.org/257407@main