[Live Range Selection] editing/selection/programmatic-selection-on-mac-is-directionless.html fails https://bugs.webkit.org/show_bug.cgi?id=248797
Reviewed by Wenson Hsieh.

Use value, selectionStart, and selectionEnd of input and textarea elements instead of relying on
getSelection().toString() to include the selected text as this will not be the case when live range
selection is enabled. Also renamed the misleadingly named "name" variable to "node" for consistency
and clarity.

* LayoutTests/editing/selection/programmatic-selection-on-mac-is-directionless.html:

Canonical link: https://commits.webkit.org/257424@main