Web Inspector: Show "Device" menu for all sessions https://bugs.webkit.org/show_bug.cgi?id=247808 rdar://102241043
Reviewed by Timothy Hatcher.

The device menu allows for applying an array of options that can negatively affect other tabs/windows if those settings
are applied globally. We currently already show the Device menu for remote inspection (like iOS), and with this patch we
will do so for local inspection (like macOS) as well, allowing developers to continue to use the rest of the browser for
normal browsing tasks, instead of using the browser's global settings for disabling core features which will break other
pages, like documentation, they are viewing.

For now we hide the User Agent dropdown, as it is not currently correctly tracking the state of the view, causing some
browser features like Safari's Responsive Design Mode to change the UA without updating this UI at the same time.

* Source/WebInspectorUI/UserInterface/Base/Main.js:
* Source/WebInspectorUI/UserInterface/Images/Computer.svg: Added.

Canonical link: https://commits.webkit.org/257426@main