Improve some CSS WPT tests so their output is easier to read and fix some mistaken expectations https://bugs.webkit.org/show_bug.cgi?id=248817 rdar://problem/103026393
Reviewed by Tim Nguyen.

* LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/cssom-getBoxQuads-001-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/cssom-getBoxQuads-001.html:
Update to be a set of separate tests instead of a single test with a lot of assertions.

* LayoutTests/imported/w3c/web-platform-tests/css/cssom/flex-serialization-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/flex-serialization.html:
Ditto.

* LayoutTests/imported/w3c/web-platform-tests/css/cssom/overflow-serialization-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/overflow-serialization.html:
Ditto.

* LayoutTests/imported/w3c/web-platform-tests/css/cssom/shorthand-values-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/shorthand-values.html:
Fixed the test to cover a case that should work where the longhand has the initial value rather
than a case that should not where the longhand has the initial keyword as its value. WebKit is now
failing this test case.

Canonical link: https://commits.webkit.org/257438@main