Disable the focus moves to an element which doesn't satisfy the spec
Disable the focus moves to an element which doesn't satisfy the spec
https://bugs.webkit.org/show_bug.cgi?id=248903

Reviewed by NOBODY (OOPS!).

This patch is to align Webkit behavior with Blink / Chromium, Gecko / Firefox and Web-Specification:

Web-Spec: https://html.spec.whatwg.org/multipage/interaction.html#sequential-focus-navigation

Merge - https://src.chromium.org/viewvc/blink?view=revision&revision=182040

This patch is to disallow behavior where if there are some elements that are
set ‘ tabindex="negative integer" ', and then, when one of them is focused on,
the next focus moves to the others. It is not aligned with web-spec.

* Source/WebCore/page/FocusController.cpp:
(FocusController::nextFocusableElementOrScopeOwner): Add 'else if' condition around trying to find first tabindex and return it
(FocusController::previousFocusableElementOrScopeOwner): Add 'else if' condition to perform similar to above
* LayoutTests/fast/events/tabindex-no-focusable.html: Add Test Case
* LayoutTests/fast/events/tabindex-no-focusable-expected.txt: Add Test Case Expectation
* LayoutTests/fast/events/tabindex-no-focusable-all-negative.html: Add Test Case
* LayoutTests/fast/events/tabindex-no-focusable-all-negative-expected.txt: Add Test Case Expectation