Shorthands are still using "initial" for longhands that are set implicitly (next batch) https://bugs.webkit.org/show_bug.cgi?id=248839 rdar://problem/103044698
Reviewed by Tim Nguyen.

Use implicit initial values instead of the implicit initial keyword. Later we will also eliminate
the concept of "implicit" and not serialize these because of their value, rather than leaving them
out of the serialization because of an implicit flag. This doesn't fix all of them. There are still
a few remaining in places that are slightly harder to remove.

* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/list-style-valid-expected.txt:
A small progression; this still does not omit enough of the properties that it should.

* Source/WebCore/css/parser/CSSPropertyParser.cpp:
(WebCore::CSSPropertyParser::consumeBorderImage): Removed unused code for parsing
-webkit-mask-box-image as a shorthand. We can restore this later if it's needed.
(WebCore::CSSPropertyParser::consumeContainerShorthand): Use initial value rather than the
initial keyword for container-type when parsing the container shorthand.
(WebCore::CSSPropertyParser::consumeListStyleShorthand): Use initial values rather than the
initial keyword for list-style-position, list-style-image, and list-style-type when parsing the
list-style shorthand.

Canonical link: https://commits.webkit.org/257556@main